+++ packages/perseus/src/widget-ai-utils/categorizer/categorizer-ai-utils.test.ts
@@ -11,7 +11,7 @@ import type {PerseusRenderer} from "../../perseus-types";
 import type {PerseusCategorizerUserInput} from "../../validation.types";
 import type {UserEvent} from "@testing-library/user-event";
 
-export const randomizedQuestion: PerseusRenderer = {
+const randomizedQuestion: PerseusRenderer = {
     content:
         "**Classify each graph according to the kind of relationship it suggests.**\n\n$\\qquad\\qquad\\quad\\text{Graph 1}\\qquad\\qquad\\quad\\qquad\\qquad\\quad\\text{Graph 2}$\n\n\n\n[[\u2603 categorizer 1]]\n\n**Graph 1.**\n\n![](https://ka-perseus-graphie.s3.amazonaws.com/049c091ed0978112aba3a36b0591d992baf7b1ac.png)\n\n**Graph 2.**\n\n![](https://ka-perseus-graphie.s3.amazonaws.com/40df186f39fb6d65de6bee0d8b681502d10cb37a.png)  \n",
     images: {},
